requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.9.3
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
Automat==24.8.1
billiard==4.2.1
cashfree-pg==4.3.10
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.2
colorama==0.4.6
constantly==23.10.4
croniter==3.0.3
cryptography==43.0.3
decouple==0.0.7
dj-rest-auth==7.0.1
Django==5.1.2
django-allauth==65.9.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-crispy-forms==2.0
django-decouple==2.1
django-extensions==3.2.3
django-filter==25.1
django-markdownx==4.0.7
django-multitenant==4.1.1
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
environ==1.0
frozenlist==1.5.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
Markdown==3.7
multidict==6.4.2
numpy==2.1.2
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.50
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.10.22
PyJWT==2.10.1
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2024.2
pyfcm==2.1.0

reportlab==4.4.3
requests==2.31.0 
sentry-sdk==1.32.0
service-identity==24.2.0
six==1.16.0
sqlparse==0.5.1
stripe==12.1.0
tomli==2.0.2
twilio==9.5.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.0.7
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.6.1
websockets==13.1
yarl==1.19.0
zope.interface==7.1.1
qrcode[pil]==7.4.2
fyers-apiv3==3.1.11
